Finished!
the base game is finished, but the ai could still make moves we don't see optimal if they still will win or tie and they are closer to the start of the tree. I hope to fix this and also try implementing alpha-beta pruning to make it a little faster even though we don't need it.